Split Ruby runner
