+++ lib/runners/shell.js
@@ -5,7 +5,7 @@ const path = require('path');
 const fs = require('fs-extra');
 const temp = require('temp').track();
 
-const ruby = require('./ruby');
+const prepareRSpec = require('./ruby/prepare-rspec');
 
 module.exports = {
   solutionOnly(opts, runCode) {
@@ -23,7 +23,7 @@ module.exports = {
       opts.fixture = `\`rm -rf /workspace/fixture.rb\` ; ${opts.fixture}`;
     }
 
-    ruby.prepareRSpec(opts, runCode);
+    prepareRSpec(opts, runCode);
   }
 };
