+++ lib/runners/sql.js
@@ -1,6 +1,7 @@
 "use strict";
 
-const ruby = require('./ruby');
+const prepareCw2 = require('./ruby/prepare-cw2');
+const prepareRSpec = require('./ruby/prepare-rspec');
 
 module.exports = {
   solutionOnly(opts, runCode) {
@@ -14,10 +15,10 @@ module.exports = {
     prepareSetup(opts);
     opts.solution = "#";
     if (opts.testFramework === 'cw-2') {
-      ruby.prepareCw2(opts, runCode);
+      prepareCw2(opts, runCode);
     }
     else {
-      ruby.prepareRSpec(opts, runCode);
+      prepareRSpec(opts, runCode);
     }
   },
   modifyOpts(opts) {
